% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_summary_plots.R
\name{ARC_URC_plot}
\alias{ARC_URC_plot}
\title{ARC_URC_plot}
\usage{
ARC_URC_plot(..., names_input = NULL, both_strand = FALSE)
}
\arguments{
\item{...}{a \code{list} of \code{ExoData} objects, or several 
\code{ExoData} objects by themselves.}

\item{names_input}{a character vector with the names to use in the
plot. If it is empty \code{ARC_URC_plot} is going to create the names
as the names of the list when they are available or is going to 
name them as Sample: 1 ,... , Sample: k.}

\item{both_strand}{A logical value indicating if the \code{DataFrame} 
contains only regions with reads aligned to both strand or all. The default
value is \code{FALSE}.}
}
\value{
A \code{ggplot2} object with the ARC vs URC plot.
}
\description{
\code{ARC_URC plot} returns a \code{ggplot} object with the ARC vs
URC plot to analyze enrichment and library complexity in ChIP-exo data.
}
\examples{
library(ChIPexoQualExample)
data(exampleExoData)
ARC_URC_plot(exampleExoData[[1]],exampleExoData[[2]])
}

