% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_summaryPlots.R
\name{ARCvURCDataFrame}
\alias{ARCvURCDataFrame}
\title{ARCvURCDataFrame}
\usage{
ARCvURCDataFrame(object, both.strand)
}
\arguments{
\item{object}{a \code{ExoData} object.}

\item{both.strand}{A logical value indicating if the \code{DataFrame} 
contains only regions with reads aligned to both strand or all. The default
value is \code{FALSE}.}
}
\value{
A \code{DataFrame} with the ARC and URC columns of the \code{ExoData}
object.
}
\description{
\code{ARCvURCDataFrame} returns a \code{DataFrame} object with the ARC and
URC columns used to plot enrichment and library complexity in ChIP-exo data.
}
\examples{
data(exoExample)
ARCvURCDataFrame(exoExample,both.strand = FALSE)
ARCvURCDataFrame(exoExample,both.strand = TRUE)
}

