% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods_ExoData.R
\docType{methods}
\name{beta1}
\alias{beta1}
\alias{beta1,ExoData-method}
\title{beta1 methods}
\usage{
beta1(object)

\S4method{beta1}{ExoData}(object)
}
\arguments{
\item{object}{a \code{ExoData} object.}
}
\value{
A numeric vector with estimated values for \eqn{\beta_1}.
}
\description{
\code{beta1} returns a vector with all the estimated values of the 
\eqn{d_i  = \beta_1 u_i + \beta_2 w_i + \epsilon_i} models fitted by 
\code{ChIPexoQual}
}
\examples{
data(exoExample)
beta1(exoExample)
}

