% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods_ExoData.R
\docType{methods}
\name{beta2}
\alias{beta2}
\alias{beta2,ExoData-method}
\title{beta2 methods}
\usage{
beta2(object)

\S4method{beta2}{ExoData}(object)
}
\arguments{
\item{object}{a \code{ExoData} object.}
}
\value{
A numeric vector with estimated values for \eqn{\beta_2}.
}
\description{
\code{beta2} returns a vector with all the estimated values of the 
\eqn{d_i  = \beta1 u_i + \beta2 w_i + \epsilon_i} models fitted by 
\code{ChIPexoQual}
}
\examples{
library(ChIPexoQualExample)
data(exampleExoData)
beta2(exampleExoData[[1]])
}

