% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{blacklists}
\alias{blacklists}
\title{\code{list} of \code{GRanges} objects with the blacklists generated by the ENCODE and 
modENCODE projects.}
\format{\code{list} of \code{GRanges} objects.}
\usage{
data(blacklists)
}
\value{
A \code{list} with the blacklists listed in https://sites.google.com/site/anshulkundaje/projects/blacklists.
}
\description{
\code{list} of \code{GRanges} objects with the blacklists generated by the ENCODE and 
modENCODE projects.
}
