% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/run_all.R
\name{bound_analysis}
\alias{bound_analysis}
\title{This functions does all the bound analysis. Briefly, it load the
reads, creates the regions and returns a collection of plots that
summarizes how the regions vary as they are filtered by depth}
\usage{
bound_analysis(exofile, mc, lowerBounds = c(1, 2, 3, 4, 5, 10, 15, 20, 25, 30,
  35, 40, 45, 50, 75, 100, 125, 150, 200, 250, 500, 750))
}
\arguments{
\item{exofile}{String that contains complete name of the reads file}

\item{mc}{An integer indicating the number of cores used}

\item{lowerBounds}{An integer vector of the bounds used in the analysis}
}
\description{
This functions does all the bound analysis. Briefly, it load the
reads, creates the regions and returns a collection of plots that
summarizes how the regions vary as they are filtered by depth
}

