% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_summary_stats.R
\name{calculate_summary}
\alias{calculate_summary}
\title{Calculate the summary statistics for each region}
\usage{
calculate_summary(region, freads, breads)
}
\arguments{
\item{region}{a IRanges object with the regions of a certain chromosome}

\item{freads}{a GRanges object with the fwd. reads of the ChIP-exo experiment}

\item{breads}{a GRanges object with the bwd. reads of the ChIP-exo experiment}
}
\value{
stats a \code{DataFrame} object with the summary statistics necessary
for the \code{ExoData}
}
\description{
Calculate the summary statistics for each region
}
\examples{
a = 1
}

