% Generated by roxygen2 (4.0.2): do not edit by hand
\name{create_regions}
\alias{create_regions}
\title{Separates a set of reads into regions in the genome}
\usage{
create_regions(reads, lower)
}
\arguments{
\item{reads,}{This is a GRanges object with the reads obtained by a ChIP-seq or a ChIP-exo experiment}

\item{lower,}{This is a positive number that indicates the lower bound used to delimite the islands}
}
\value{
A set of regions in the genome
}
\description{
Separates a set of reads into regions in the genome
}

