% Generated by roxygen2 (4.0.2): do not edit by hand
\name{data.table.GRanges}
\alias{data.table.GRanges}
\title{Function to convert to data.table}
\usage{
data.table.GRanges(x)
}
\arguments{
\item{x,}{A GRanges object}
}
\value{
A data.table object
}
\description{
Function to convert to data.table
}

