% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/summary_statistics.R
\name{depth_by_region}
\alias{depth_by_region}
\title{Calculates the read depth of a collection of regions in the genome}
\usage{
depth_by_region(extract_reads, mcores)
}
\arguments{
\item{extract_reads,}{A data.table object based in a GRanges format, width an
  additional column named regions which partition the reads}

\item{mcores,}{A numeric value indicating the number of cores to use}
}
\description{
Calculates the read depth of a collection of regions in the genome
}

