% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/summary_statistics.R
\name{depth_from_reads}
\alias{depth_from_reads}
\title{This function calculates the number of reads that overlaps a set of regions}
\usage{
depth_from_reads(chr_reads, chr_nregions = NULL)
}
\arguments{
\item{chr_reads}{A data.table containing the reads that overlaps the set of regions, with a region key}

\item{chr_nregions,}{An integer with the expected size of the vector, if it is NULL is going to define as the possible number of regions in chr_reads}
}
\value{
A numeric vector with the depth for each region
}
\description{
This function calculates the number of reads that overlaps a set of regions
}

