% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleExoData}
\alias{exampleExoData}
\title{\code{ExoData} results for FoxA1 ChIP-exo experiment}
\format{ExoData object, which are \code{GRanges} with additional columns.}
\usage{
data(exampleExoData)
}
\description{
List of \code{ExoData} objects, generated with \code{ChIPexoQual} and the
files:
}
\details{
\itemize{
\item ChIPexo_carroll_FoxA1_mouse_rep1_chr1.bam
\item ChIPexo_carroll_FoxA1_mouse_rep2_chr1.bam
\item ChIPexo_carroll_FoxA1_mouse_rep3_chr1.bam
}
}

