% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/summary_plots.R
\name{filter_label_plot}
\alias{filter_label_plot}
\title{Returns a ggplot object with an area plot of the label distribution per lower bound}
\usage{
filter_label_plot(lowerBounds, chr_depths, labels)
}
\arguments{
\item{lowerBounds}{An integer vector with the lower bounds used to filter}

\item{chr_depths}{A list of vectors with the islands depths per chromosome}

\item{labels}{A list of character vectors with the labels for each island}
}
\description{
Returns a ggplot object with an area plot of the label distribution per lower bound
}

