% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/summary_plots.R
\name{filter_MA_plot}
\alias{filter_MA_plot}
\title{Returns a ggplot object with a collection of MA plots faceted by lower bound}
\usage{
filter_MA_plot(lowerBounds, chr_depths, chr_M_values, chr_A_values, mc,
  smooth = FALSE)
}
\arguments{
\item{lowerBounds}{An integer vector with the lower bounds used to filter}

\item{chr_depths}{A list of vectors with the islands depths per chromosome}

\item{chr_M_values}{A list of vectors with the M coordinate of an MA plot}

\item{chr_A_values}{A list of vectors with the A coordinate of an MA plot}

\item{mc}{Number of cores used}
}
\description{
Returns a ggplot object with a collection of MA plots faceted by lower bound
}

