% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/summary_plots.R
\name{filter_scatter_plot}
\alias{filter_scatter_plot}
\title{Returns a ggplot object with a collection of scatter plots faceted by lower bound. It filters the labels by the islands such that the depth is greater than lowerBound}
\usage{
filter_scatter_plot(lowerBounds, var1, var2, filtered_summary, mc,
  smooth = FALSE)
}
\arguments{
\item{lowerBounds}{An integer vector with the lower bounds used to filter}

\item{var1}{String with the name of which variable is going to be plotted in the x-axis}

\item{var2}{String with the name of which variable is going to be plotted in the y-axis}

\item{filtered_summary}{A list of data.tables with the already filtered summary statistics}

\item{mc}{Number of cores used}

\item{smooth}{A boolean variable indicating if a loess regression line is going to be
added to the plot}
}
\description{
Returns a ggplot object with a collection of scatter plots faceted by lower bound. It filters the labels by the islands such that the depth is greater than lowerBound
}

