% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_summaryPlots.R
\name{filterQuantiles}
\alias{filterQuantiles}
\title{Calculates the FSR distribution of all regions with depth > value
 
Calculates the FSR distribution of all regions with depth > value}
\usage{
filterQuantiles(value, DF, quantiles)
}
\arguments{
\item{value}{a numeric value with the depth lower bound.}

\item{DF}{a \code{DataFrame} with depth and FSR.}

\item{quantiles}{a numeric vector with the quantiles used to summarize
the FSR distribution.}
}
\value{
a \code{DataFrame} with three columns depth, quantiles and FSR.
}
\description{
Calculates the FSR distribution of all regions with depth > value
 
Calculates the FSR distribution of all regions with depth > value
}

