% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_summaryPlots.R
\name{filterRegionComp}
\alias{filterRegionComp}
\title{Calculates the region composition probabilities for all 
regions with depth > value
 
Calculates the region composition probabilities for all 
regions with depth > value}
\usage{
filterRegionComp(value, DF)
}
\arguments{
\item{value}{a numeric value with the depth lower bound.}

\item{DF}{a \code{DataFrame} with depth and FSR.}
}
\value{
a \code{DataFrame} with three columns d, label and prob.
}
\description{
Calculates the region composition probabilities for all 
regions with depth > value
 
Calculates the region composition probabilities for all 
regions with depth > value
}

