% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_summaryPlots.R
\name{FSRDistDataFrame}
\alias{FSRDistDataFrame}
\title{FSRDistDataFrame}
\usage{
FSRDistDataFrame(object, quantiles, depth.values, both.strand)
}
\arguments{
\item{object}{a \code{ExoData} object.}

\item{quantiles}{a numeric vector with the quantiles used to estimate the
FSR distribution at a given depth.}

\item{depth.values}{a numeric vector indicating the regions with depth 
less or equal to, that are going to be filtered out.}

\item{both.strand}{a logical value indicating if the \code{DataFrame} 
contains only regions with reads aligned to both strand or all.}
}
\value{
A \code{DataFrame} object with the FSR distribution of the 
\code{ExoData} object.
}
\description{
\code{FSRDistDataFrame} returns a \code{DataFrame} object with the Forward 
Strand Ratio distribution of the \code{ExoData} object to analyze strand 
imbalance in ChIP-exo data.
}
\examples{
data(exoExample)
FSRDistDataFrame(exoExample,quantiles = c(.25,.5,.75),
  depth.values = seq_len(25),both.strand = FALSE)
}

