% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_summary_plots.R
\name{MA_plot}
\alias{MA_plot}
\title{MA_plot}
\usage{
MA_plot(..., names_input = NULL)
}
\arguments{
\item{...}{a \code{list} of \code{ExoData} objects, or several 
\code{ExoData} objects by themselves.}

\item{names_input}{a character vector with the names to use in the
plot. If it is empty \code{ARC_URC_plot} is going to create the names
as the names of the list when they are available or is going to 
name them as Sample: 1 ,... , Sample: k.}
}
\value{
A \code{ggplot2} object with the MA plot.
}
\description{
\code{MA_plot} returns a \code{ggplot} object with the MA plot to 
analyze the strand imbalance in ChIP-exo data.
}
\examples{
library(ChIPexoQualExample)
data(exampleExoData)
MA_plot(exampleExoData)
MA_plot(exampleExoData[[1]],exampleExoData[[2]])
}

