% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/summary_statistics.R
\name{MA_values}
\alias{MA_values}
\title{Update the summary_table including the MA-plot values}
\usage{
MA_values(summary_table)
}
\arguments{
\item{summary_table}{A data.table with summary statistics}
}
\value{
Updated summary table including M and A values
}
\description{
Update the summary_table including the MA-plot values
}

