% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_summaryPlots.R
\name{MADataFrame}
\alias{MADataFrame}
\title{MADataFrame}
\usage{
MADataFrame(object)
}
\arguments{
\item{object}{A \code{ExoData} object.}
}
\value{
A \code{DataFrame} with two columns: M and A.
}
\description{
Returns a \code{DataFrame} with the info. to generate a MA
plot to analyze the strand imbalance in ChIP-exo data.
}
\examples{
data(exoExample)
MADataFrame(exoExample)
}

