% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_summaryPlots.R
\name{MAplot}
\alias{MAplot}
\title{MAplot}
\usage{
MAplot(..., names.input = NULL)
}
\arguments{
\item{...}{a \code{list} of \code{ExoData} objects, or several 
\code{ExoData} objects by themselves.}

\item{names.input}{a character vector with the names to use in the
plot. If it is empty \code{MAplot} is going to create the names
as the names of the list when they are available or is going to 
name them as Sample: 1 ,... , Sample: k.}
}
\value{
A \code{ggplot2} object with the MA plot.
}
\description{
\code{MAplot} returns a \code{ggplot} object with the MA plot to 
analyze the strand imbalance in ChIP-exo data.
}
\examples{
data(exoExample)
MAplot(exoExample)
}

