% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/summary_statistics.R
\name{mcshiftapply}
\alias{mcshiftapply}
\title{Returns a list (or vector when simplify=TRUE) in the following manner:
Let ‘i’ be the indices in 'SHIFT', ‘X_i = window(X, 1, length(X) - SHIFT[i])’,
and ‘Y_i = window(Y, 1 + SHIFT[i], length(Y) )’. Calculates the
set of ‘FUN(X_i, Y_i, ...)’ values and return the results in a convenient form}
\usage{
mcshiftapply(SHIFT, FUN, X, Y, mc, ..., simplify = TRUE)
}
\arguments{
\item{SHIFT}{A non-negative integer vector of shift values.}

\item{FUN}{The function, found via ‘match.fun’, to be applied to each set of
shifted vectors.}

\item{X}{The Vector or R vector objects to shift.}

\item{Y}{The Vector or R vector objects to shift.}

\item{...}{Further arguments for FUN}

\item{simplify}{A logical value specifying whether or not the result should be
simplified to a vector}
}
\description{
Returns a list (or vector when simplify=TRUE) in the following manner:
Let ‘i’ be the indices in 'SHIFT', ‘X_i = window(X, 1, length(X) - SHIFT[i])’,
and ‘Y_i = window(Y, 1 + SHIFT[i], length(Y) )’. Calculates the
set of ‘FUN(X_i, Y_i, ...)’ values and return the results in a convenient form
}

