% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/summary_statistics.R
\name{npos_by_region}
\alias{npos_by_region}
\title{Calculates the number of unique reads for a collection of regions in the genome}
\usage{
npos_by_region(extract_reads, st, mcores)
}
\arguments{
\item{extract_reads,}{A data.table object based in a GRanges format, width an
  additional column named regions which partition the reads}

\item{st,}{A character with "+" for forward and "-" for reverse fragment reads}

\item{mcores,}{A numeric value indicating the number of cores to use}
}
\description{
Calculates the number of unique reads for a collection of regions in the genome
}

