% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods_ExoData.R
\docType{methods}
\name{nreads}
\alias{nreads}
\alias{nreads,ExoData-method}
\title{nreads methods}
\usage{
nreads(object)

\S4method{nreads}{ExoData}(object)
}
\arguments{
\item{object}{A \code{ExoData} object.}
}
\value{
The number of reads in the \code{ExoData} object.
}
\description{
\code{nreads} returns the number of reads in the object.
}
\examples{
data(exoExample)
nreads(exoExample)
}

