% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods_ExoData.R
\docType{methods}
\name{param_dist}
\alias{param_dist}
\alias{param_dist,ExoData-method}
\title{param_dist methods}
\usage{
param_dist(object)

\S4method{param_dist}{ExoData}(object = "ExoData")
}
\arguments{
\item{object}{a \code{ExoData} object.}
}
\value{
A \code{data.table} with the fitted values of \eqn{\beta_1} and 
\eqn{\beta_2}.
}
\description{
\code{param_dist} returns a \code{data.table} with all the estimated 
coefficients in the \eqn{d_i  = \beta_1 u_i + \beta_2 w_i + \epsilon_i} models 
fitted by \code{ChIPexoQual}
}
\examples{
a = 1
}

