% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_summary_plots.R
\name{param_dist_boxplot}
\alias{param_dist_boxplot}
\title{param_dist_boxplot}
\usage{
param_dist_boxplot(..., names_input = NULL, which_param = "beta1")
}
\arguments{
\item{...}{a \code{list} of \code{ExoData} objects, or several 
\code{ExoData} objects by themselves.}

\item{names_input}{a character vector with the names to use in the
plot. If it is empty \code{param_dist_boxplot} is going to create the 
names as the names of the list when they are available or is going to 
name them as Sample: 1 ,... , Sample: k.}

\item{which_param}{a character value with either \code{"beta1"} or 
\code{"beta2"} that determines which paramters in the model 
d_i ~ u_i + w_i to plot. The default value is \code{"beta1"}.}
}
\value{
A \code{ggplot2} object with the boxplot of the chosen 
parameter
}
\description{
\code{param_dist_boxplot} returns a \code{ggplot} object with a 
boxplot comparing the \code{ntimes} estimations of the chosen 
parameter.
}
\examples{
library(ChIPexoQualExample)
data(exampleExoData)
param_dist_boxplot(exampleExoData[[1]],exampleExoData[[2]])
}

