% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods_ExoData.R
\docType{methods}
\name{paramDist}
\alias{paramDist}
\alias{paramDist,ExoData-method}
\title{paramDist methods}
\usage{
paramDist(object)

\S4method{paramDist}{ExoData}(object = "ExoData")
}
\arguments{
\item{object}{a \code{ExoData} object.}
}
\value{
A \code{DataFrame} with the fitted values of \eqn{\beta_1} and 
\eqn{\beta_2}.
}
\description{
\code{paramDist} returns a \code{DataFrame} with all the estimated 
coefficients in the \eqn{d_i  = \beta_1 u_i + \beta_2 w_i + \epsilon_i} models 
fitted by \code{ChIPexoQual}
}
\examples{
data(exoExample)
paramDist(exoExample)
}

