% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_summaryPlots.R
\name{paramDistBoxplot}
\alias{paramDistBoxplot}
\title{paramDistBoxplot}
\usage{
paramDistBoxplot(..., names.input = NULL, which.param = "beta1",
  sort.as.numeric = FALSE)
}
\arguments{
\item{...}{a \code{list} of \code{ExoData} objects, or several 
\code{ExoData} objects by themselves.}

\item{names.input}{a character vector with the names to use in the
plot. If it is empty \code{paramDistBoxplot} is going to create the 
names as the names of the list when they are available or is going to 
name them as Sample: 1 ,... , Sample: k.}

\item{which.param}{a character value with either \code{"beta1"} or 
\code{"beta2"} that determines which paramters in the model 
depth_i ~ uniquePos_i + width_i to plot. The default value is 
\code{"beta1"}.}

\item{sort.as.numeric}{a logical value indicating if the values of 
\code{names.input} are meant to be interpreted as numeric and sorted 
accordingly.}
}
\value{
A \code{ggplot2} object with the boxplot of the chosen 
parameter
}
\description{
\code{paramDistBoxplot} returns a \code{ggplot} object with a 
boxplot comparing the \code{ntimes} estimations of the chosen 
parameter.
}
\examples{
data(exoExample)
paramDistBoxplot(exoExample)
}

