% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/summary_statistics.R
\name{pbc}
\alias{pbc}
\title{Calculate the PCR bottleneck coefficient for a data set}
\usage{
pbc(reads_table, mcores)
}
\arguments{
\item{reads_table}{List of data.table with the experiment's separated reads by each region}

\item{mcores}{Numeric value with the number of cores to use}
}
\description{
Calculate the PCR bottleneck coefficient for a data set
}

