% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/summary_plots.R
\name{plot_cover}
\alias{plot_cover}
\title{Returns a ggplot object with a region's ChIP (chip, seq or exo) profile. This plot
colours both strands as separate, and in particular considers negative count for the
reverse strand fragments.}
\usage{
plot_cover(fwd_cover, bwd_cover, region_start, region_end, depth,
  annot = NULL, chr, ext)
}
\arguments{
\item{fwd_cover}{This is an Rle object built with the \link{\code{coverage}} function
from IRanges, considering only the forward strand reads}

\item{bwd_cover}{This is an Rle object built with the \link{\code{coverage}} function
from IRanges, considering only the reverse strand reads}

\item{region_start}{This is the genomic coordinate where the region starts}

\item{region_end}{This is the genomic coordinate where the region ends}

\item{depth}{This is the depth of the experiment to normalize the coverage's as if
they were a million reads in the experiment}

\item{annot}{The summary statistics that are going to be written in the plot for
comparison purposes. If annot == NULL, then doesn't write anything}

\item{ext}{A number with the number of bases used to extend the plot to each side}
}
\description{
Returns a ggplot object with a region's ChIP (chip, seq or exo) profile. This plot
colours both strands as separate, and in particular considers negative count for the
reverse strand fragments.
}

