% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/summary_plots.R
\name{positions_reads_map}
\alias{positions_reads_map}
\title{Returns a ggplot object with a map of the number of positions vs the number of reads, where the color denotes how many regions agree with that specific combination}
\usage{
positions_reads_map(summary_table, mp = Inf, md = Inf)
}
\arguments{
\item{summary_table}{A data.table with the summary statistics per region containing at least the npos and depth fields}

\item{mp}{Max number of positions}

\item{md}{Max depth}
}
\description{
Returns a ggplot object with a map of the number of positions vs the number of reads, where the color denotes how many regions agree with that specific combination
}

