% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/create_regions.R
\name{reads_overlaps}
\alias{reads_overlaps}
\title{Find the overlaps between the reads and regions for certain chromosome}
\usage{
reads_overlaps(chr, chr_islands, chr_reads)
}
\arguments{
\item{chr,}{This is the chromosome (or genome) for which we are looking the regions}

\item{chr_islands,}{IRanges object with a collection of regions in certain chromosome (or genome)}

\item{chr_reads,}{GRanges object with a set of reads for certain chromosome (or genome)}
}
\value{
A list with all the overlaps between reads and island for each chromosome
}
\description{
Find the overlaps between the reads and regions for certain chromosome
}

