% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_summaryPlots.R
\name{regionCompDataFrame}
\alias{regionCompDataFrame}
\title{regionCompDataFrame}
\usage{
regionCompDataFrame(object, depth.values)
}
\arguments{
\item{object}{a \code{ExoData} object.}

\item{depth.values}{a numeric vector indicating the regions with depth 
less or equal to, that are going to be filtered out. The defaulta values 
are \code{seq_len(50)}.}
}
\value{
A \code{DataFrame} with three columns: depth, label and propotion
of regions at a given depth level with the respective label.
}
\description{
\code{regionCompDataFrame} returns a \code{DataFrame} with the info. 
necessary to generate the Region Composition plot to analyze strand
imbalance in ChIP-exo data.
}
\examples{
data(exoExample)
regionCompDataFrame(exoExample,seq_len(10))
}

