% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/summary_statistics.R
\name{summary_statistics}
\alias{summary_statistics}
\title{Update the summary_table with summary statistics}
\usage{
summary_statistics(summary_table)
}
\arguments{
\item{summary_table}{A data.table with the fields width, fwd_npos, bwd_npos, fwd_depth and bwd_depth}
}
\value{
Updated summary_table with compound statistics
}
\description{
Update the summary_table with summary statistics
}

