% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/create_regions.R
\name{create_regions}
\alias{create_regions}
\title{Separates a set of reads into regions in the genome}
\usage{
create_regions(reads, lower, rangesOnly = TRUE)
}
\arguments{
\item{reads,}{This is a GRanges object with the reads obtained by a ChIP-seq or a ChIP-exo experiment}

\item{lower,}{This is a positive number that indicates the lower bound used to delimite the islands}

\item{rangesOnly,}{boolean to asses is we want also the coverage to be calculated. The default value if true}
}
\value{
A set of regions in the genome
}
\description{
Separates a set of reads into regions in the genome
}

