% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{ExoData-class}
\alias{ExoData}
\alias{ExoData-class}
\title{ExoData object and constructors}
\usage{
ExoData(file, height = 1, mc.cores = getOption("mc.cores", 2L),
  save_reads = FALSE, nregions = 1000, ntimes = 1000)
}
\arguments{
\item{file}{a character value with location of the bam file with the aligned
reads.}

\item{height}{a numeric value indicating the value used to slice the coverage
of the experiment into a set of regions.}

\item{mc.cores}{a numeric value with the number of cores to use,
i.e. at most how many child processes will be run simultaneously.}

\item{save_reads}{a logical value to indicate if the reads are stored in the
\code{ExoData} object. The default value is \code{FALSE}.}

\item{nregions}{a numeric value indicating the number of regions sampled to 
estimate the quality parameter distributions. The default value is 1e3.}

\item{ntimes}{a numeric value indicating the number of times that regions are 
sampled to estimate the quality parameter distributions. The default value
is 1e3.}
}
\value{
\code{ExoData} returns a \code{ExoData} object which contains the
aggregated coverage of the experiment, the set of islands and a collection
of summary statistics used to asses the quality of a ChIP-exo/nexus sample.
}
\description{
\code{ExoData} is a subclass of \code{GenomicRanges}, used to asses the 
quality of ChIP-exo/nexus sample.
}
\examples{

a = 1

}

