% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exoExample}
\alias{exoExample}
\title{\code{ExoData} results for FoxA1 ChIP-exo experiment}
\format{ExoData object, which are \code{GRanges} with additional columns.}
\usage{
data(exoExample)
}
\value{
An \code{ExoData} object with the 3rd replicate of the FoxA1 
experiment from \code{ChIPExoQualExample}.
}
\description{
\code{ExoData} object, generated with \code{ChIPexoQual} and the
file:
}
\details{
\itemize{
\item ChIPexo_carroll_FoxA1_mouse_rep3_chr1.bam
}
}
