% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/extract_reads.R
\name{extract_reads}
\alias{extract_reads}
\title{Extract the reads that overlap an island from the whole reads list for a certain chromosome}
\usage{
extract_reads(chr, chr_reads, chr_overlaps, chr_islands)
}
\arguments{
\item{chr,}{This is the chromosome for which we are looking the regions}

\item{chr_reads,}{GRanges object with a set of reads for certain chromosome}

\item{chr_overlaps,}{A Hits object with all the overlaps between the reads and the islands or certain chromosome}

\item{chr_islands,}{IRanges object with a collection of regions in certain chromosome}
}
\value{
A list with all the overlaps between reads and island for each chromosome
}
\description{
Extract the reads that overlap an island from the whole reads list for a certain chromosome
}

