% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/summary_plots.R
\name{filter_regions_plot}
\alias{filter_regions_plot}
\title{Returns a ggplot object with a boxplot of lowerBounds vs chr_measurement. It filters the measurement by the islands such that the depth is greater than lowerBound}
\usage{
filter_regions_plot(lowerBounds, filtered_summary, measure_var, measure_label,
  mc, log = FALSE)
}
\arguments{
\item{lowerBounds}{An integer vector with the lower bounds used to filter}

\item{filtered_summary}{A list of data.tables with the already filtered summary statistics}

\item{measure_var}{A string with the name of the variable to be drawn}

\item{measure_label}{A string with the y-axis label}

\item{mc}{Number of cores used}

\item{log}{A boolean variable indicating if the y-axis is going to be rescaled with log10}
}
\description{
Returns a ggplot object with a boxplot of lowerBounds vs chr_measurement. It filters the measurement by the islands such that the depth is greater than lowerBound
}

